[CSL-818] WIP - Decouple derive package for LTS-8 upgrade
derive was removed from stackage 8.1. To upgrade to LTS-8 we need to move all
derive-based code into test executables, rather than the main library.

This is a WIP checkpoint, and may not prove to be the right direction. First, I
commented out a lot of the code that depends on derive. Next I found avvmmigrate
has arbitrary instances that depend on the commented out derive calls, so
somehow those modules need to be made accessible to it. I started making a new
cardano-sl-arbitrary library package, but may need to take a different approach.